requirements.txt
aiohttp==3.8.4
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-debug-toolbar==4.0.0
django-ninja==0.21.0
django-ratelimit==4.0.0
django-seed==0.3.1
django-user-agents==0.4.0
djangorestframework==3.14.0
Faker==18.5.1
geoip2==4.6.0
gunicorn==20.1.0
idna==3.4
maxminddb==2.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
pathspec==0.11.1
psycopg2==2.9.6
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.5.0
ua-parser==0.16.1
urllib3==1.26.6
user-agents==2.2.0
yarl==1.9.2
